Adding all buttons
