+++ DeCo19/src/components/AlertaAglo/style.js
@@ -2,7 +2,6 @@ import { StyleSheet } from 'react-native';
 
 const styles = StyleSheet.create({
 
-
    alertaTitle: {
       fontSize: 18,
       fontFamily: 'Roboto',
@@ -23,17 +22,6 @@ const styles = StyleSheet.create({
       borderRadius: 3,
    },
 
-   buttonBox: {
-      flex: 1,
-      padding: 15,
-      justifyContent: 'flex-end',
-      margin: 15,
-      height: 50,
-      width: 80,
-      backgroundColor: '#ff0000',
-      borderRadius: 3,
-   },
-
    submitButton: {
       backgroundColor: '#005BAB',
       padding: 10,
@@ -47,7 +35,6 @@ const styles = StyleSheet.create({
       fontSize: 16,
       flex: 1,
       padding: 10,
-      justifyContent: 'flex',
       textAlign: 'center',
       color: 'white',
       fontFamily: 'Roboto',
